% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft_cache.R
\name{dft_cache_path}
\alias{dft_cache_path}
\title{Get drift cache directory path}
\usage{
dft_cache_path(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Override the default cache location.
If NULL, uses \code{rappdirs::user_cache_dir("drift")}.}
}
\value{
Character path to the cache directory.
}
\description{
Returns the path to the drift tile cache directory. Creates it if
it doesn't exist.
}
\examples{
dft_cache_path()

}
