% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift-package.R
\docType{package}
\name{drift-package}
\alias{drift}
\alias{drift-package}
\title{drift: Detecting Riparian and Inland Floodplain Transitions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fetch STAC-hosted land cover rasters, classify and summarize change over time within user-provided AOI polygons, and produce standardized maps, plots, and tables for reporting. Generic STAC pipeline works with any classified raster (Esri IO LULC, ESA WorldCover, custom COGs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/drift}
  \item \url{https://newgraphenvironment.github.io/drift/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/drift/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item New Graph Environment [copyright holder]
}

}
\keyword{internal}
