% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.drift.R
\name{sim.drift}
\alias{sim.drift}
\title{Simulate Genetic Drift}
\usage{
sim.drift(n.pop,n,freq,n.gen,n.loci,n.time,s.plot,s.table)
}
\arguments{
\item{n.pop}{Number of effective population}

\item{n}{Ploidy: Haploid=1, Diploid=2...}

\item{freq}{initial frequency of allele (0-1)}

\item{n.gen}{Number of generations to simulate Default:100}

\item{n.loci}{Number of genes/loci tested}

\item{n.time}{Number of simulations to perform}

\item{s.plot}{Save the image of plot? (TRUE/FALSE) Default:FALSE}

\item{s.table}{Save the table of numerical valuest? (TRUE/FALSE) Default:FALSE}
}
\value{
Frequency Plots. If save is true, image and tabel text file are saved in working dierctory.
}
\description{
This function allows you to simulate the change in allelic frequencies due to random drift in a Wright-Fisher model
}
\examples{
sim.drift(30,2,0.25,100,10,3,TRUE,FALSE)
sim.drift(30,2,0.25,100,10,3)
}
